+++ src/components/Record/Card/Card.js
@@ -1,15 +1,15 @@
 import "./Card.scss";
 import { HiPencil } from "react-icons/hi";
-import { Link, useNavigate } from 'react-router-dom';
-
-function Card({ applicationId, emoji, title, status, endDate, numberOfReplies }) {
-
-  const navigate = useNavigate()
-  const today = new Date();
-  const dday = new Date(endDate);
-  const gap = dday.getTime() - today.getTime();
-  const result = Math.ceil(gap / (1000 * 60 * 60 * 24));
-
+import { Link } from "react-router-dom";
+import { statusCheck } from "../../../utils/etc/StatusDivision";
+function Card({
+  applicationId,
+  emoji,
+  title,
+  status,
+  endDate,
+  numberOfReplies,
+}) {
   return (
     <Link className="record-card" to={`/application/${applicationId}/manage`}>
       <div className="record-card-com">
@@ -19,20 +19,12 @@ function Card({ applicationId, emoji, title, status, endDate, numberOfReplies })
         </div>
         <div className="record-card-bottom">
           <p className="record-card-bottom-p">
-            {status === "ALWAYS" ? (
-              <span>상시</span>
-            ) : status === "NOT_STARTED" ? (
-              <span>시작 전</span>
-            ) : status === "IN_PROGRESS" ? (
-              <span>{result}일 남음</span>
-            ) : (
-              <span>완료됨</span>
-            )}
+            {statusCheck(status, endDate)}
             <span className="record-card-bar">|</span>
             <span>{numberOfReplies}개의 신청</span>
           </p>
           <Link to={`/application/${applicationId}/update`}>
-            <HiPencil className="record-card-bottom-pen"/>
+            <HiPencil className="record-card-bottom-pen" />
           </Link>
         </div>
       </div>
