+++ src/index.scss
@@ -21,6 +21,7 @@ body {
   color: black;
   width: 100vw;
   height: calc(var(--vh, 1vh) * 100);
+  overflow-x: hidden;
 }
 
 ul {
